+++ src/main/java/com/tms/shop/controllers/ProductController.java
@@ -17,7 +17,7 @@ public class ProductController {
     private final ProductService productService;
 
     @GetMapping("/{id}")
-    public ModelAndView openProductPage(@PathVariable int id) {
+    public ModelAndView openProductPage(@PathVariable Long id) {
         return productService.getProductData(id);
     }
 }
